+++ spec/autoscaler/counter_cache_redis_spec.rb
@@ -41,7 +41,7 @@ describe Autoscaler::CounterCacheRedis do
   end
 
   it 'passed a plain connection' do
-    connection = Redis.connect(:url => 'redis://localhost:9736', :namespace => 'autoscaler')
+    connection = Redis.connect(:url => 'redis://localhost:9736')
     cache = cut.new connection
     cache.counter = 5
     expect(cache.counter).to eq 5